WIP: refactoring changed storage.go
